sandbox_pms_mvp/requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Migrate>=4.0,<5.0
psycopg[binary]>=3.2,<4.0
argon2-cffi>=23.1,<24.0
cryptography>=44.0,<45.0
pyotp>=2.9,<3.0
pytest>=8.3,<9.0
gunicorn>=22.0,<23.0
